+++ test/surgex/parser/parsers/string_parser_test.exs
@@ -46,6 +46,19 @@ defmodule Surgex.Parser.StringParserTest do
     assert StringParser.call("  abcde  ", opt) == {:error, :too_long}
   end
 
+  test "regex" do
+    opt = [regex: ~r/^[a-h]{1,4}$/]
+    assert StringParser.call("", opt) == {:error, :bad_format}
+    assert StringParser.call("abc", opt) == {:ok, "abc"}
+    assert StringParser.call("axn", opt) == {:error, :bad_format}
+    assert StringParser.call("abcde", opt) == {:error, :bad_format}
+  end
+
+  test "regex check is done after trimming" do
+    opt = [regex: ~r/^[a-h]{1,4}$/, trim: true]
+    assert StringParser.call("  abc ", opt) == {:ok, "abc"}
+  end
+
   test "unsupported input type" do
     assert StringParser.call(1.4) == {:error, :invalid_string}
   end